requirement.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
opentelemetry-api==1.28.1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-flask==0.49b1
opentelemetry-instrumentation-sqlalchemy==0.49b1
opentelemetry-instrumentation-wsgi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
packaging==24.2
protobuf==5.28.3
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
thrift==0.21.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.16.0
zipp==3.21.0
